requirements.txt
Django==1.3.1
django-tastypie==0.9.11
mimeparse==0.1.3
